move molescules dir
