+++ __test__/molecules/forms/EditableProperty.test.tsx
@@ -1,5 +1,5 @@
 import { mount, shallow, ReactWrapper } from 'enzyme';
-import { EditableProperty } from '../../../src/molecules/forms/EditableProperty';
+import { EditableProperty } from '../../../src/components/molecules/forms/EditableProperty';
 import { isVisible } from '../../util';
 import { act } from 'react-dom/test-utils';
